Online Chess
Written in Java with GWT (Google Web Toolkit) for compiling to native
JavaScript. Onlinedemo at:
http://mt111047.students.fhstp.ac.at/WebChess/war/
(->At least two players have to join the same Session. The name of the
session is freely selectable. Further guests can join the session for
watching a game. This game also supports a P2P Videochat [powered by
WebRTC]. Many game sessions can be started and played at the same time.
Each new session has to have a new name). The server is NodeJS-based and
is included in the directory "ChessServer".